add SVG tester output
