+++ apps/owidbot/chart_diff.py
@@ -1,78 +1,59 @@
-from apps.staging_sync.cli import _get_container_name
+import pandas as pd
+from sqlmodel import Session
+
+from apps.staging_sync.cli import _get_container_name, _get_engine_for_env, _modified_chart_ids_by_admin
+from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
 
 
 def run(branch: str) -> str:
     container_name = _get_container_name(branch) if branch else "dry-run"
-    chart_diff = "TODO"
+
+    chart_diff = format_chart_diff(call_chart_diff(branch))
 
     body = f"""
 <details>
-<summary><b>chart-diff</b>: </summary>
+<summary><b>Chart diff</b>: </summary>
 {chart_diff}
-[Details](http://{container_name}/etl/wizard/Chart%20Diff)
+<a href="http://{container_name}/etl/wizard/Chart%20Diff">Details</a>
 </details>
     """.strip()
 
     return body
 
 
-# Uncomment the code once the ChartDiffModified class is implemented
+def call_chart_diff(branch: str) -> pd.DataFrame:
+    source_engine = _get_engine_for_env(branch)
+    # TODO: this should be live
+    target_engine = _get_engine_for_env("master")
 
-# import pandas as pd
-# from sqlmodel import Session
+    df = []
+    with Session(source_engine) as source_session:
+        with Session(target_engine) as target_session:
+            modified_chart_ids = _modified_chart_ids_by_admin(source_session)
 
-# from apps.staging_sync.cli import _get_container_name, _get_engine_for_env, _modified_chart_ids_by_admin
-# from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
+            for chart_id in modified_chart_ids:
+                diff = ChartDiffModified.from_chart_id(chart_id, source_session, target_session)
+                df.append(
+                    {
+                        "chart_id": diff.chart_id,
+                        "approved": diff.approved,
+                        "is_new": diff.is_new,
+                    }
+                )
 
-
-# def run(branch: str):
-#     container_name = _get_container_name(branch) if branch else "dry-run"
-
-#     chart_diff = format_chart_diff(call_chart_diff(branch))
-
-#     body = f"""
-# <details>
-# <summary><b>Chart diff</b>: </summary>
-# {chart_diff}
-# <a href="http://{container_name}/etl/wizard/Chart%20Diff">Details</a>
-# </details>
-#     """.strip()
-#     return body
-
-
-# def call_chart_diff(branch: str) -> pd.DataFrame:
-#     source_engine = _get_engine_for_env(branch)
-#     # TODO: this should be live
-#     target_engine = _get_engine_for_env("master")
-
-#     df = []
-#     with Session(source_engine) as source_session:
-#         with Session(target_engine) as target_session:
-#             modified_chart_ids = _modified_chart_ids_by_admin(source_session)
-
-#             for chart_id in modified_chart_ids:
-#                 diff = ChartDiffModified.from_chart_id(chart_id, source_session, target_session)
-#                 df.append(
-#                     {
-#                         "chart_id": diff.chart_id,
-#                         "approved": diff.approved,
-#                         "is_new": diff.is_new,
-#                     }
-#                 )
-
-#     return pd.DataFrame(df)
+    return pd.DataFrame(df)
 
 
-# def format_chart_diff(df: pd.DataFrame) -> str:
-#     if df.empty:
-#         return "No new or modified charts."
+def format_chart_diff(df: pd.DataFrame) -> str:
+    if df.empty:
+        return "No new or modified charts."
 
-#     new = df[df.is_new]
-#     modified = df[~df.is_new]
+    new = df[df.is_new]
+    modified = df[~df.is_new]
 
-#     return f"""
-# <ul>
-#     <li>{len(new)} new charts ({new.approved.sum()} approved)</li>
-#     <li>{len(modified)} modified charts ({modified.approved.sum()} approved)</li>
-# </ul>
-#     """.strip()
+    return f"""
+<ul>
+    <li>{len(new)} new charts ({new.approved.sum()} approved)</li>
+    <li>{len(modified)} modified charts ({modified.approved.sum()} approved)</li>
+</ul>
+    """.strip()
